refactor(graph): move `resolve_commits`
