+++ src/util.rs
@@ -2,9 +2,6 @@
 
 use std::path::PathBuf;
 
-use fn_error_context::context;
-use git2::ErrorCode;
-
 /// Returns a path for a given file, searching through PATH to find it.
 pub fn get_from_path(exe_name: &str) -> Option<PathBuf> {
     std::env::var_os("PATH").and_then(|paths| {
@@ -39,46 +36,4 @@ pub fn get_sh() -> Option<PathBuf> {
         }
     }
     get_from_path(exe_name)
-}
-
-/// The result of attempting to resolve commits.
-pub enum ResolveCommitsResult<'repo> {
-    /// All commits were successfully resolved.
-    Ok {
-        /// The commits.
-        commits: Vec<git2::Commit<'repo>>,
-    },
-
-    /// The first commit which couldn't be resolved.
-    CommitNotFound {
-        /// The identifier of the commit, as provided by the user.
-        commit: String,
-    },
-}
-
-/// Parse strings which refer to commits, such as:
-///
-/// - Full OIDs.
-/// - Short OIDs.
-/// - Reference names.
-#[context("Resolving commits")]
-pub fn resolve_commits(
-    repo: &git2::Repository,
-    hashes: Vec<String>,
-) -> anyhow::Result<ResolveCommitsResult> {
-    let mut commits = Vec::new();
-    for hash in hashes {
-        let commit = match repo.revparse_single(&hash) {
-            Ok(commit) => match commit.into_commit() {
-                Ok(commit) => commit,
-                Err(_) => return Ok(ResolveCommitsResult::CommitNotFound { commit: hash }),
-            },
-            Err(err) if err.code() == ErrorCode::NotFound => {
-                return Ok(ResolveCommitsResult::CommitNotFound { commit: hash })
-            }
-            Err(err) => return Err(err.into()),
-        };
-        commits.push(commit)
-    }
-    Ok(ResolveCommitsResult::Ok { commits })
 }
